+++ CLightSource.h
@@ -26,7 +26,7 @@ public :
 
 	virtual Vector3 EvalAmbient(Vector3 _material_Ka) = 0;
 	virtual Vector3 EvalDiffuse(Vector3 _material_kd, Vector3 _N, Vector3 _L) = 0;
-	virtual Vector3 EvalSpecular(Vector3 _material_Ks, double _material_Shininess, Vector3 _N, Vector3 _L, Vector3 _V) = 0;
+	virtual Vector3 EvalSpecular(Vector3 _material_Ks, float _material_Shininess, Vector3 _N, Vector3 _L, Vector3 _V) = 0;
 };
 
 #endif
